+++ src/main/java/it/gesev/dao/TipoDerrateDAO.java
@@ -8,8 +8,8 @@ import it.gesev.entities.TipoDerrata;
 public interface TipoDerrateDAO 
 {
 	public List<TipoDerrata> leggiTuttiTipiDerrate();
-	public long createTipoDerrata(TipoDerrata tipoDerrata);
-	public long deleteTipoDerrata(long codiceTipoDerrata);
-	public long updateTipoDerrata(int codiceTipoDerrata, TipoDerrata tipoDerrata);
+	public int createTipoDerrata(TipoDerrata tipoDerrata);
+	public int deleteTipoDerrata(int codiceTipoDerrata);
+	public int updateTipoDerrata(int codiceTipoDerrata, TipoDerrata tipoDerrata);
 	public List<TipoDerrata> cercaTipoDerrataConColonna(Map<String, String> mappa);
 }
